asthma_backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
flask-jwt-extended==4.6.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
bcrypt==4.1.3
numpy==1.26.4
tensorflow==2.16.2
librosa==0.10.2
scikit-learn==1.5.1
